chore: restructure packages
