+++ src/main/kotlin/fr/nicopico/n2rss/analytics/AnalyticsConfig.kt
@@ -18,7 +18,10 @@
 
 package fr.nicopico.n2rss.analytics
 
-import fr.nicopico.n2rss.analytics.simpleanalytics.SimpleAnalyticsService
+import fr.nicopico.n2rss.analytics.service.AnalyticsService
+import fr.nicopico.n2rss.analytics.service.AnalyticsServiceDelegate
+import fr.nicopico.n2rss.analytics.service.NoOpAnalyticsService
+import fr.nicopico.n2rss.analytics.service.simpleanalytics.SimpleAnalyticsService
 import fr.nicopico.n2rss.config.N2RssProperties
 import org.slf4j.LoggerFactory
 import org.springframework.context.annotation.Bean
