+++ src/main/kotlin/fr/nicopico/n2rss/analytics/models/AnalyticsEvent.kt
@@ -15,7 +15,7 @@
  * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  */
-package fr.nicopico.n2rss.analytics
+package fr.nicopico.n2rss.analytics.models
 
 sealed class AnalyticsEvent {
     /**
