+++ src/main/kotlin/fr/nicopico/n2rss/analytics/service/AnalyticsService.kt
@@ -15,30 +15,12 @@
  * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  */
+package fr.nicopico.n2rss.analytics.service
 
-package fr.nicopico.n2rss.analytics
+import fr.nicopico.n2rss.analytics.models.AnalyticsEvent
+import fr.nicopico.n2rss.analytics.models.AnalyticsException
 
-import org.jetbrains.annotations.VisibleForTesting
-
-class AnalyticsServiceDelegate(
-    @VisibleForTesting
-    internal val analyticsServices: List<AnalyticsService>,
-): AnalyticsService {
-
-    override fun track(event: AnalyticsEvent) {
-        val results = analyticsServices.map {
-            runCatching { it.track(event) }
-        }
-        if (results.any { it.isFailure }) {
-            val throwable = AnalyticsException("Some analytics failed to process the event $event")
-            results
-                .mapNotNull { it.exceptionOrNull() }
-                .forEach { throwable.addSuppressed(it) }
-
-            throw AnalyticsException(
-                "Some analytics failed to process the event $event",
-                throwable,
-            )
-        }
-    }
+interface AnalyticsService {
+    @Throws(AnalyticsException::class)
+    fun track(event: AnalyticsEvent)
 }
