+++ src/main/kotlin/fr/nicopico/n2rss/analytics/service/data/DataAnalyticsService.kt
@@ -15,11 +15,11 @@
  * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  */
-package fr.nicopico.n2rss.analytics.data
+package fr.nicopico.n2rss.analytics.service.data
 
-import fr.nicopico.n2rss.analytics.AnalyticsEvent
-import fr.nicopico.n2rss.analytics.AnalyticsException
-import fr.nicopico.n2rss.analytics.AnalyticsService
+import fr.nicopico.n2rss.analytics.models.AnalyticsEvent
+import fr.nicopico.n2rss.analytics.models.AnalyticsException
+import fr.nicopico.n2rss.analytics.service.AnalyticsService
 import fr.nicopico.n2rss.config.N2RssProperties
 import kotlinx.datetime.Clock
 import org.slf4j.LoggerFactory
