+++ src/main/kotlin/fr/nicopico/n2rss/mail/EmailChecker.kt
@@ -17,10 +17,10 @@
  */
 package fr.nicopico.n2rss.mail
 
-import fr.nicopico.n2rss.config.NewsletterConfiguration
 import fr.nicopico.n2rss.mail.client.EmailClient
 import fr.nicopico.n2rss.mail.models.Email
 import fr.nicopico.n2rss.monitoring.MonitoringService
+import fr.nicopico.n2rss.newsletter.NewsletterConfiguration
 import fr.nicopico.n2rss.newsletter.data.PublicationRepository
 import fr.nicopico.n2rss.newsletter.handlers.NewsletterHandler
 import fr.nicopico.n2rss.newsletter.handlers.process
